Rename Queue Processor -> Listener
Renames Queue Processor to Listener to better reflect what it is
actually doing
